+++ src/wolff_producers_sup.erl
@@ -18,7 +18,7 @@
 
 -export([start_link/0, init/1]).
 
--export([ensure_present/3, ensure_absence/2]).
+-export([ensure_present/3, ensure_absence/1]).
 
 -define(SUPERVISOR, ?MODULE).
 
@@ -36,8 +36,8 @@ init([]) ->
 %% ensure a client started under supervisor
 -spec ensure_present(wolff:client_id(), wolff_producers:id(), wolff_producers:config()) ->
   {ok, pid()} | {error, term()}.
-ensure_present(ClientId, ID, Config) ->
-  ChildSpec = child_spec(ClientId, ID, Config),
+ensure_present(ClientId, ProducerId, Config) ->
+  ChildSpec = child_spec(ClientId, ProducerId, Config),
   case supervisor:start_child(?SUPERVISOR, ChildSpec) of
     {ok, Pid} -> {ok, Pid};
     {error, {already_started, Pid}} -> {ok, Pid};
@@ -45,23 +45,19 @@ ensure_present(ClientId, ID, Config) ->
   end.
 
 %% ensure client stopped and deleted under supervisor
--spec ensure_absence(wolff:client_id(), wolff_producers:id()) -> ok.
-ensure_absence(ClientId, ID) ->
-  Id = worker_id(ClientId, ID),
-  case supervisor:terminate_child(?SUPERVISOR, Id) of
+-spec ensure_absence(wolff_producers:id()) -> ok.
+ensure_absence(ProducerId) ->
+  case supervisor:terminate_child(?SUPERVISOR, ProducerId) of
     ok ->
-      ok = wolff_producers:cleanup_workers_table(ClientId, ID),
-      ok = supervisor:delete_child(?SUPERVISOR, Id);
+      ok = wolff_producers:cleanup_workers_table(ProducerId),
+      ok = supervisor:delete_child(?SUPERVISOR, ProducerId);
     {error, not_found} ->
       ok
   end.
 
-child_spec(ClientId, ID, Config) ->
-  #{id => worker_id(ClientId, ID),
-    start => {wolff_producers, start_link, [ClientId, ID, Config]},
+child_spec(ClientId, ProducerId, Config) ->
+  #{id => ProducerId,
+    start => {wolff_producers, start_link, [ClientId, ProducerId, Config]},
     restart => transient,
     type => worker
    }.
-
-worker_id(ClientId, ID) ->
-  {ClientId, ID}.